requirements.txt
torch>=1.2
torchvision
numpy
pillow
scikit-learn